requirements.txt
Django==1.7.1

markdown

django-tinymce

Pillow

django-password-reset

